# docs/labos-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vyoma LabOS",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/branches": {
      "get": {
        "tags": [
          "branches"
        ],
        "summary": "Get Branches",
        "description": "List all branches of the current organization.",
        "operationId": "get_branches_api_v1_branches_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BranchResponse"
                  },
                  "type": "array",
                  "title": "Response Get Branches Api V1 Branches Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "branches"
        ],
        "summary": "Create Branch",
        "description": "Create a new branch (admin role required).",
        "operationId": "create_branch_api_v1_branches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/branches/availability": {
      "get": {
        "tags": [
          "branches"
        ],
        "summary": "Get Branches Availability",
        "description": "Get detailed branch/location availability including operating hours, services, and contact info.",
        "operationId": "get_branches_availability_api_v1_branches_availability_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Get Branches Availability Api V1 Branches Availability Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/bookings/doctor": {
      "post": {
        "tags": [
          "bookings"
        ],
        "summary": "Book Doctor Appointment",
        "description": "Book a doctor appointment for a patient (Capability 9).",
        "operationId": "book_doctor_appointment_api_v1_bookings_doctor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorAppointmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Book Doctor Appointment Api V1 Bookings Doctor Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/bookings/lab": {
      "post": {
        "tags": [
          "bookings"
        ],
        "summary": "Book Lab Appointment",
        "description": "Book a lab appointment or home sample collection request (Capability 10).",
        "operationId": "book_lab_appointment_api_v1_bookings_lab_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabBookingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Book Lab Appointment Api V1 Bookings Lab Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/customer-care/handoff": {
      "post": {
        "tags": [
          "customer-care"
        ],
        "summary": "Create Customer Care Handoff",
        "description": "Create a customer care escalation/handoff ticket (Capability 11).",
        "operationId": "create_customer_care_handoff_api_v1_customer_care_handoff_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCareHandoffCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Customer Care Handoff Api V1 Customer Care Handoff Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "customer-care"
        ],
        "summary": "List Customer Care Tickets",
        "description": "List customer-care handoff tickets with optional status/priority filtering.",
        "operationId": "list_customer_care_tickets_api_v1_customer_care_handoff_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Customer Care Tickets Api V1 Customer Care Handoff Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BranchCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Name"
          },
          "code": {
            "type": "string",
            "maxLength": 20,
            "minLength": 2,
            "title": "Code"
          },
          "address": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "contact_info": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Info"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "code",
          "organization_id"
        ],
        "title": "BranchCreate"
      },
      "BranchResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Name"
          },
          "code": {
            "type": "string",
            "maxLength": 20,
            "minLength": 2,
            "title": "Code"
          },
          "address": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "contact_info": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Info"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "code",
          "id",
          "organization_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "BranchResponse"
      },
...
      "CustomerCareHandoffCreate": {
        "properties": {
          "organization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "channel": {
            "type": "string",
            "title": "Channel",
            "default": "whatsapp"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "report_query"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "normal"
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "title": "Summary"
          },
          "conversation_transcript": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Transcript"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "CustomerCareHandoffCreate"
      },
...
      "DoctorAppointmentCreate": {
        "properties": {
          "organization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "patient_id": {
            "type": "integer",
            "title": "Patient Id"
          },
          "doctor_id": {
            "type": "integer",
            "title": "Doctor Id"
          },
          "slot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Id"
          },
          "branch_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Id"
          },
          "appointment_date": {
            "type": "string",
            "format": "date",
            "title": "Appointment Date"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "consultation_type": {
            "type": "string",
            "title": "Consultation Type",
            "default": "in_person"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "doctor_id",
          "appointment_date",
          "start_time",
          "end_time"
        ],
        "title": "DoctorAppointmentCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabBookingCreate": {
        "properties": {
          "organization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "patient_id": {
            "type": "integer",
            "title": "Patient Id"
          },
          "branch_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Id"
          },
          "booking_type": {
            "type": "string",
            "title": "Booking Type",
            "default": "home_collection"
          },
          "preferred_date": {
            "type": "string",
            "format": "date",
            "title": "Preferred Date"
          },
          "preferred_slot": {
            "type": "string",
            "title": "Preferred Slot"
          },
          "address": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "tests_requested": {
            "items": {},
            "type": "array",
            "title": "Tests Requested"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "preferred_date",
          "preferred_slot",
          "tests_requested"
        ],
        "title": "LabBookingCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}